pyproject.toml
[project]
name = "nonebot-plugin-saaweather"
version = "0.1.0"
description = "nonebot-plugin-saaweather"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot-plugin-send-anything-anywhere>=0.5.0",
    "nonebot-plugin-orm>=0.6.3",
    "nonebot-plugin-htmlrender>=0.2.3",
    "nonebot2>=2.1.3",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
builtin_plugins = ["echo"]

[tool.pdm.dev-dependencies]
dev = [
    "nonebug>=0.3.5",
    "nonebug-saa>=0.3.1",
    "pytest>=7.4.4",
    "respx>=0.20.2",
    "pytest-asyncio>=0.23.4",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
addopts = "--cov=nonebot --cov-append --cov-report=term-missing"
filterwarnings = ["error", "ignore::DeprecationWarning"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["nonebot", "tests"]
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

line-length = 88
target-version = "py310"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./",
  ] },
  { root = "./" },
]

typeCheckingMode = "basic"
reportShadowedImports = false
disableBytesTypePromotions = true
